requirements.txt
Flask==3.0.0
line-bot-sdk==3.5.0
supabase==2.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
APScheduler==3.10.4
Pillow==10.2.0
python-dateutil==2.8.2
flask-cors==4.0.0
python-jose==3.3.0
cryptography==41.0.7
